Characteristics of RAPELD plots in Ilha Grande, RJ
Characteristics of RAPELD plots in Ilha Grande, RJ. The vegetation type follows the classification by Veloso et al. (1991) e Brazilian Institute of Geography and Statistics (IBGE, 2012), successional stages are in accordance with the National Council for the Environment (Brasil 1994) and the other characteristics correspond to the average values of each plot followed by the standard deviation in meters (m) and centimeters (cm).